02-Back-End/dengue_api-bkp-1/pyproject.toml
[tool.poetry]
name = "dengue_api"
version = "0.1.0"
description = ""
authors = ["andreperna <andreperna@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.78.0"}
uvicorn = "^0.17.6"
mysql-connector-python = "^8.0.29"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
